# Vectorized HttpUserAgentParser.TryExtractVersion