@gfoidl 2025-09-06
@BenjaminAbt we added tests for the faulty user agent that caused troubles the last time? I don't remember it...