@gfoidl 2025-10-05
Re https://github.com/mycsharp/HttpUserAgentParser/pull/79#discussion_r2332896727:
ATM that approach has the problem, that it doesn't honor the order of some bots that we rely on, so some tests are failing. I need to try to come up with a solution for this. When I can't find one, I'll still create a PR, so that the approach is documented (because basically I like this one, as it only scans the input once, instead of multiple times). But I don't know how fast I can find a solution.

Thus I'm going to merge this PR now.